+++ GraphicsPlayground/GraphicsPlayground/Engine/Source/Runtime/Engine/Classes/Components/StaticMeshComponent.cpp
@@ -170,7 +170,7 @@ int UStaticMeshComponent::CheckRayIntersection(const FVector& InRayOrigin, const
         return 0;
     }
     
-    const TArray<UINT>& Indices = RenderData->Indices;
+    const TArray<uint16>& Indices = RenderData->Indices;
     const int32 IndexNum = Indices.Num();
     const bool bHasIndices = (IndexNum > 0);
